[REFACTOR] Home에서 사용되는 기능을 HomeService로 옮기고, ProfileService에서 쓰이지 않는 기능 제거
