+++ src/main/java/uni/backend/controller/HomeController.java
@@ -1,7 +1,9 @@
 package uni.backend.controller;
 
+import java.awt.print.Pageable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import uni.backend.domain.dto.HomeDataResponse;
 import uni.backend.domain.dto.HomeProfileResponse;
+import uni.backend.service.HomeService;
 import uni.backend.service.ProfileService;
 
 import java.util.List;
@@ -20,12 +23,14 @@ import java.util.List;
 public class HomeController {
 
     @Autowired
-    ProfileService profileService;
+    HomeService homeService;
 
 
     @GetMapping("/home")
-    public ResponseEntity<HomeDataResponse> getHomeProfiles() {
-        HomeDataResponse homeDataResponse = profileService.getHomeDataProfiles();
+    public ResponseEntity<HomeDataResponse> getHomeProfiles(
+        @PageableDefault(size = 8) Pageable pageable) {
+
+        HomeDataResponse homeDataResponse = homeService.getHomeDataProfiles(pageable);
 
         return ResponseEntity.ok(homeDataResponse);
     }
