+++ src/main/java/uni/backend/domain/Profile.java
@@ -1,6 +1,8 @@
 package uni.backend.domain;
 
 import jakarta.persistence.*;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -14,65 +16,77 @@ import java.util.List;
 @Setter
 public class Profile {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "profile_id")
-  private Integer profileId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "profile_id")
+    private Integer profileId;
 
-  @OneToOne
-  @JoinColumn(name = "user_id", nullable = true) // null 허용
-  private User user;
+    @OneToOne
+    @JoinColumn(name = "user_id", nullable = true) // null 허용
+    private User user;
 
-  @Column(nullable = true) // null 허용
-  private LocalDateTime createdAt;
+    @Column(nullable = true) // null 허용
+    private LocalDateTime createdAt;
 
-  @Column(nullable = true) // null 허용
-  private LocalDateTime updatedAt;
+    @Column(nullable = true) // null 허용
+    private LocalDateTime updatedAt;
 
-  @Column(nullable = true) // null 허용
-  private String imgProf;
+    @Column(nullable = true) // null 허용
+    private String imgProf;
 
-  @Column(nullable = true) // null 허용
-  private String imgBack;
+    @Column(nullable = true) // null 허용
+    private String imgBack;
 
-  @Column(nullable = true) // null 허용
-  private String region;
+    @Column(nullable = true) // null 허용
+    private String region;
 
-  @Column(nullable = true) // null 허용
-  private String description;
+    @Column(nullable = true) // null 허용
+    private String description;
 
-  @Column(nullable = true) // null 허용
-  private Integer numEmployment;
+    @Column(nullable = true) // null 허용
+    private Integer numEmployment;
 
-  @Column(nullable = true) // null 허용
-  private Double star;
+    @Column(nullable = true) // null 허용
+    private Double star;
 
-  @Column(nullable = true) // null 허용
-  private String time;
+    @Column(nullable = true) // null 허용
+    private String time;
 
-  @OneToMany(mappedBy = "profile", cascade = CascadeType.ALL, orphanRemoval = true)
-  private List<MainCategory> mainCategories = new ArrayList<>();
+    @OneToMany(mappedBy = "profile", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<MainCategory> mainCategories = new ArrayList<>();
 
-  public void addMainCategory(MainCategory mainCategory) {
-    this.mainCategories.add(mainCategory);
-    mainCategory.setProfile(this);
-  }
+    public void addMainCategory(MainCategory mainCategory) {
+        this.mainCategories.add(mainCategory);
+        mainCategory.setProfile(this);
+    }
 
-  public void removeMainCategory(MainCategory mainCategory) {
-    mainCategories.remove(mainCategory);
-    mainCategory.setProfile(null);
-  }
+    public void removeMainCategory(MainCategory mainCategory) {
+        mainCategories.remove(mainCategory);
+        mainCategory.setProfile(null);
+    }
 
-  @PrePersist
-  public void prePersist() {
-    this.createdAt = LocalDateTime.now();
-    this.numEmployment = 0;
-    this.star = 0.0;
-  }
+    @PrePersist
+    public void prePersist() {
+        this.createdAt = LocalDateTime.now();
+        this.numEmployment = 0;
+        this.star = 0.0;
+    }
 
 
-  @PreUpdate
-  public void preUpdate() {
-    this.updatedAt = LocalDateTime.now();
-  }
+    @PreUpdate
+    public void preUpdate() {
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    public List<String> getHashtagStringList() {
+        List<String> hashtags;
+        hashtags = this.mainCategories.stream()
+            .map(mainCategory -> {
+                Hashtag hashtag = mainCategory.getHashtag();
+                return hashtag != null ? hashtag.getHashtagName() : null;
+            })
+            .filter(Objects::nonNull)
+            .collect(Collectors.toList());
+        return hashtags;
+    }
 }
